Improve version comparison with subversions
Subversions in CPEs are some times handled differently by the NVD. For
example "cpe:2.3:a:proftpd:proftpd:1.3.5b:-:*:..." vs.
"cpe:2.3:a:proftpd:proftpd:1.3.3:c:*:...". Similarly, CVE-2019-19269 has
version_end of "1.3.3e", which resulted in a positive match before for
"cpe:2.3:a:proftpd:proftpd:1.3.5:f:*:...".

Furthermore, when search_vulns creates CPEs, it forces the user to
choose the one where the subversion is placed correctly. This update
helps here as well and is able to compensate when the user chose the
wrong one.