Major refactoring and interface and code changes
This update makes search_vulns modular and refactors most of the code.
As a result, it's probably best to do a fresh install and data update.
Furthermore, external tools may have to be adjusted to the small changes
in function signatures, the data format or changes in the API.

The big advantage with this update is that new features and data sources
can be added to the search engine way more easily now.